gmm_training/requirements.txt
tensorflow
numpy
scikit-learn
Pillow
